change location of useTracking hook, rename useCurrentTrackStore
