+++ src/frontend/Navigation/ScreenGroups/TabBar/TrackingTabBarIcon.tsx
@@ -1,7 +1,7 @@
 import React, {FC} from 'react';
 import {StyleSheet, View} from 'react-native';
 import {TabBarIcon} from './TabBarIcon';
-import {useTracking} from '../../../hooks/tracks/useTracking';
+import {useTracking} from '../../../hooks/useTracking.ts';
 import {Text} from '../../../sharedComponents/Text';
 import {TabBarIconProps} from '../../types';
 import {useTrackTimerContext} from '../../../contexts/TrackTimerContext';
