+++ src/frontend/hooks/useTracking.ts
@@ -1,10 +1,10 @@
 import * as Location from 'expo-location';
 import * as TaskManager from 'expo-task-manager';
 import {useCallback, useState} from 'react';
-import {useCurrentTrackStore} from './useCurrentTrackStore';
+import {usePersistedTrack} from './persistedState/usePersistedTrack.ts';
 import React from 'react';
-import {FullLocationData} from '../../sharedTypes/location';
-import {useGPSModalContext} from '../../contexts/GPSModalContext';
+import {FullLocationData} from '../sharedTypes/location.ts';
+import {useGPSModalContext} from '../contexts/GPSModalContext.tsx';
 
 export const LOCATION_TASK_NAME = 'background-location-task';
 
@@ -16,9 +16,9 @@ type LocationCallbackInfo = {
 export function useTracking() {
   const {bottomSheetRef} = useGPSModalContext();
   const [loading, setLoading] = useState(false);
-  const addNewLocations = useCurrentTrackStore(state => state.addNewLocations);
-  const setTracking = useCurrentTrackStore(state => state.setTracking);
-  const isTracking = useCurrentTrackStore(state => state.isTracking);
+  const addNewLocations = usePersistedTrack(state => state.addNewLocations);
+  const setTracking = usePersistedTrack(state => state.setTracking);
+  const isTracking = usePersistedTrack(state => state.isTracking);
 
   React.useEffect(() => {
     TaskManager.defineTask(
