+++ src/frontend/screens/MapScreen/ObsevationMapLayer.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import MapboxGL from '@rnmapbox/maps';
 import {useAllObservations} from '../../hooks/useAllObservations';
 import {useNavigationFromHomeTabs} from '../../hooks/useNavigationWithTypes';
-import {useCurrentTrackStore} from '../../hooks/tracks/useCurrentTrackStore';
+import {usePersistedTrack} from '../../hooks/persistedState/usePersistedTrack.ts';
 
 const DEFAULT_MARKER_COLOR = '#F29D4B';
 
@@ -17,7 +17,7 @@ const layerStyles = {
 export const ObservationMapLayer = () => {
   const observations = useAllObservations();
   const {navigate} = useNavigationFromHomeTabs();
-  const isTracking = useCurrentTrackStore(state => state.isTracking);
+  const isTracking = usePersistedTrack(state => state.isTracking);
   const featureCollection: GeoJSON.FeatureCollection = {
     type: 'FeatureCollection',
     features: mapObservationsToFeatures(observations),
