+++ src/frontend/screens/MapScreen/track/SaveTrackScreen.tsx
@@ -19,7 +19,7 @@ import {useBottomSheetModal} from '../../../sharedComponents/BottomSheetModal';
 import DiscardIcon from '../../../images/delete.svg';
 import ErrorIcon from '../../../images/Error.svg';
 import {TabName} from '../../../Navigation/types.ts';
-import {useCurrentTrackStore} from '../../../hooks/tracks/useCurrentTrackStore.ts';
+import {usePersistedTrack} from '../../../hooks/persistedState/usePersistedTrack.ts';
 import {useNavigationFromHomeTabs} from '../../../hooks/useNavigationWithTypes.ts';
 import {useFocusEffect} from '@react-navigation/native';
 import {SaveTrackButton} from './saveTrack/SaveTrackButton.tsx';
@@ -27,9 +27,7 @@ import Close from '../../../images/close.svg';
 
 export const SaveTrackScreen = () => {
   const navigation = useNavigationFromHomeTabs();
-  const clearCurrentTrack = useCurrentTrackStore(
-    state => state.clearCurrentTrack,
-  );
+  const clearCurrentTrack = usePersistedTrack(state => state.clearCurrentTrack);
   const {formatMessage: t} = useIntl();
   const [description, setDescription] = useState('');
   const {sheetRef, isOpen, openSheet, closeSheet} = useBottomSheetModal({
