+++ src/frontend/screens/ObservationEdit/SaveButton.tsx
@@ -12,7 +12,7 @@ import {UIActivityIndicator} from 'react-native-indicators';
 import {useCreateBlobMutation} from '../../hooks/server/media';
 import {DraftPhoto, Photo} from '../../contexts/PhotoPromiseContext/types';
 import {useDraftObservation} from '../../hooks/useDraftObservation';
-import {useCurrentTrackStore} from '../../hooks/tracks/useCurrentTrackStore';
+import {usePersistedTrack} from '../../hooks/persistedState/usePersistedTrack.ts';
 import SaveCheck from '../../images/CheckMark.svg';
 
 const m = defineMessages({
@@ -74,10 +74,8 @@ export const SaveButton = ({
   const createObservationMutation = useCreateObservation();
   const editObservationMutation = useEditObservation();
   const createBlobMutation = useCreateBlobMutation();
-  const addNewTrackLocation = useCurrentTrackStore(
-    state => state.addNewLocations,
-  );
-  const addNewTrackObservation = useCurrentTrackStore(
+  const addNewTrackLocation = usePersistedTrack(state => state.addNewLocations);
+  const addNewTrackObservation = usePersistedTrack(
     state => state.addNewObservation,
   );
